% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca.R
\name{baixar_pag_busca}
\alias{baixar_pag_busca}
\title{Baixar páginas da busca}
\usage{
baixar_pag_busca(q, dir, t = 1)
}
\description{
Baixa todas as páginas que contém as listas de resultados da sua busca.
}
\note{
É importante baixar os dados html para se precisar rodar novamente o
código, você não ter que baixá-los novamente.
}

