% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticias.R
\name{baixar_urls}
\alias{baixar_urls}
\title{Baixar lista de urls}
\usage{
baixar_urls(url, dir, t = 1)
}
\arguments{
\item{url}{vetor com as urls que deseja salvar}

\item{dir}{diretório em que deseja salvar as pastas}

\item{t}{tempo entre o download das páginas}
}
\description{
Baixar lista de urls
}

