% Generated by roxygen2 (4.0.1): do not edit by hand
\name{busca_noticias}
\alias{busca_noticias}
\title{Busca Notícias}
\usage{
busca_noticias(datai, dataf)
}
\arguments{
\item{datai}{data de início da busca}

\item{dataf}{data de fim da busca}
}
\description{
Busca as notícias do site da folha por datas específicas
}

