% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticias.R
\name{check_download}
\alias{check_download}
\title{Checar download}
\usage{
check_download(f)
}
\arguments{
\item{f}{path para o arquivo baixado}
}
\description{
Checa se o downlaod foi concluído com sucesso.
A única condição atualmente é que o arquivo possua pelo menos 1.8kB
}

