% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticias.R
\name{procurar_refresh}
\alias{procurar_refresh}
\title{Procurar refresh}
\usage{
procurar_refresh(path)
}
\description{
Essa função serve apenas para procurar a url do refresh.
Algumas notícias foram movidas e se encontram em outra URL.
}

