% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca.R
\name{qtd_resultados}
\alias{qtd_resultados}
\title{Quantidade de resultados}
\usage{
qtd_resultados(p)
}
\arguments{
\item{p}{pagina recebida pelo httr::GET}
}
\value{
o número de resultados encontrados na busca
}
\description{
Quantidade de resultados
}

