% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca.R
\name{query}
\alias{query}
\title{Query}
\usage{
query(p, sd = "", ed = "")
}
\arguments{
\item{p}{palavras chave separadas por espaço}

\item{sd}{data de inicio da busca}

\item{ed}{data de fim da busca}
}
\description{
Cria a query
}
\note{
as datas devem ser passadas no formato dia-mes-ano
}

