% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scrap_noticia}
\alias{scrap_noticia}
\title{Scrap Noticia}
\usage{
scrap_noticia(url)
}
\arguments{
\item{url}{url da notícia do site da folha que você deseja obter os dados (caracter)}
}
\description{
Faz webscrap de uma notícia do site da folha a partir de seu url
}

