% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scrap_noticias}
\alias{scrap_noticias}
\title{Scrap Notícias}
\usage{
scrap_noticias(urls)
}
\arguments{
\item{urls}{urls das notícias das quais você deseja obter os dados (vetor de caracteres)}
}
\description{
Faz webscrap de varias noticias do site da folha a partir de um vetor de urls
}

