% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca.R
\name{tabelar_busca}
\alias{tabelar_busca}
\title{Tabelar p치gians de busca}
\usage{
tabelar_busca(dir)
}
\arguments{
\item{dir}{diret칩rio com os arquivos das p치ginas de buscas baixadas.}
}
\description{
Tabelar p치gians de busca
}

