% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticias.R
\name{processar_noticias}
\alias{processar_noticias}
\title{Processar notícias}
\usage{
processar_noticias(dir)
}
\arguments{
\item{dir}{diretorio com todos os arquivos das notícias}
}
\value{
retorna um vetor de chr com os textos de cada uma das notícias.
}
\description{
Processar notícias
}

